+++ src/resource/RefraxResource.js
@@ -6,9 +6,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 const RefraxConstants = require('RefraxConstants');
+const RefraxTools = require('RefraxTools');
 const RefraxOptions = require('RefraxOptions');
 const RefraxResourceBase = require('RefraxResourceBase');
 const STATUS_STALE = RefraxConstants.status.STALE;
+const STATUS_COMPLETE = RefraxConstants.status.COMPLETE;
 const TIMESTAMP_LOADING = RefraxConstants.timestamp.loading;
 const CLASSIFY_ITEM = RefraxConstants.classify.item;
 // WeakMap offers a ~743% performance boost (~0.55ms => ~0.074ms) per fetch
@@ -58,9 +60,9 @@ class RefraxResource extends RefraxResourceBase {
 
       // NOTE: we invalidate before potentially subscribing
       if (this._options.invalidate) {
-        // shortcut for no options
+        // `invalidate: true` is an alias for `{ noPropgate: true }`
         if (this._options.invalidate === true) {
-          this._options.invalidate = {noPropagate: true};
+          this._options.invalidate = { noPropagate: true };
         }
 
         this.invalidate(this._options.invalidate);
@@ -87,13 +89,22 @@ class RefraxResource extends RefraxResourceBase {
 
       subscriber();
 
+      // 'touch' actions that originate from ourself come from `_fetchFragment` so we can
+      // safely ignore them as that implicitly updates our cache state
+      if (event.action === 'touch' && event.invoker === this) {
+        return;
+      }
+
       // If we are an item resource and we encounter a destroy event, we switch on the
       // 'no fetching' option so we can still passively poll the data but not cause a re-fetch
       if (descriptor.classify === CLASSIFY_ITEM && event.action === 'destroy') {
         this._options.noFetchGet = true;
       }
 
-      this._updateCache();
+      this._updateCache({
+        noPropagate: !!event.noPropagate,
+        noFetchGet: !!event.noFetchGet
+      });
 
       if (event.noPropagate !== true) {
         this.emit('change', this, event);
@@ -103,15 +114,22 @@ class RefraxResource extends RefraxResourceBase {
     subscriber();
   }
 
-  _fetchFragment() {
-    return this.fetch({ fragmentOnly: true });
+  _fetchFragment(options = {}) {
+    const fragment = this.fetch(RefraxTools.extend({}, options, {
+      fragmentOnly: true
+    }));
+
+    ResourceMap.set(this, fragment || {});
+
+    return fragment;
   }
 
-  _updateCache() {
-    const fragment = this._fetchFragment();
+  _updateCache(options = {}) {
+    const fragment = this._fetchFragment(options);
 
-    this._dispatchLoad && this._dispatchLoad(fragment && fragment.data);
-    ResourceMap.set(this, fragment);
+    if (this._dispatchLoad && fragment && fragment.status === STATUS_COMPLETE) {
+      this._dispatchLoad(fragment.data);
+    }
   }
 
   invalidate(options = {}) {
@@ -119,7 +137,9 @@ class RefraxResource extends RefraxResourceBase {
 
     this._generateDescriptor(null, descriptorOptions, (descriptor) => {
       if (descriptor.store) {
-        descriptor.store.invalidate(descriptor, options);
+        descriptor.store.invalidate(descriptor, RefraxTools.extend({}, options, {
+          invoker: this
+        }));
       }
 
       if (options.cascade === true) {